Use proxy.ts, not middleware.ts, for dev CORS
Next 16 renamed the file. On 16.3.1 the old name still loads and runs, but
every matched API route then answers 404 whatever the handler returns, even a
no-op that returns undefined. That reads as a missing route rather than a
misplaced file, so the reason is written down in the file itself.

Also widens the trusted Expo origins across the ports Expo walks up through
when 8081 is busy, since a fixed port goes stale the first time two dev
servers overlap and sign-in then fails for a reason nothing explains.
